Add files via upload
According to a 2019 industry forecast by Meticulous Research, the healthcare IT market is expected to grow at 13.8% (Compound Annual Growth Rate) between 2019 and 2027, to reach a value of $511 billion1. On a similar note, we witness that the global open source services market is expected to grow from $11.40 billion in 2017 to $32.95 billion by 2022, at a Compound Annual Growth Rate (CAGR) of 23.65% during the forecast period (2017 – 2022)2. If we were to compare the rapid growth rates of both these industries with the rate at which the open source medical software industry is growing, we would find that the open source medical software industry will only account for 10% of the growing healthcare IT market. In light of this alarming fact, we are led to question what factors have contributed to the comparatively stunted growth of the open source medical software industry. This paper examines these factors and evaluates the opportunity cost associated with choosing closed source medical software over open source medical software.

The most significant barriers affecting the use of open source medical software include the role of software vendors in facilitating the adoption of open source software in hospitals, the perceived lack of security of patient data, misconceptions about the quality, and accountability of open source products, public funding of proprietary software projects and the lack of programming knowledge that exists among medical professionals.